feat: add warning in inactive step block for mismatched options
